services/detection-engine/requirements.txt
neo4j==5.14.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1
python-json-logger==2.0.7
schedule==1.2.0
kafka-python==2.0.2
